Virtual Stock Market
Virtual Stock Market is a desktop program that allows users to create their accounts. With that
account, they are able to construct their portfolios and buy/sell stocks. Values of stocks will
be changed according to users’ actions. Changes in values of stocks will be presented. Users

will be able to follow their stock values on the screen. Users also can see their portfolio histo-
ry and compare it with the other users’.